GetDirectoryContents() should properly retrieve all directories now.

+++ User_Cleanup/Program.cs
@@ -5,6 +5,8 @@ using System.Threading.Tasks;
 using System.Windows.Forms;
 using Microsoft.Win32;
 using System.IO;
+using System.Security.AccessControl;
+
 namespace User_Cleanup
 {
     static class Program
@@ -13,21 +15,25 @@ namespace User_Cleanup
         // before anything can be deleted.
         // Look for copycat profles. e.g. "[username].[DOMAIN]"
 
-        public static string profilesPath = @"c:\Users\";
-        public static string root = "HKEY_LOCAL_MACHINE";
-        public static string keyName = @"ProfileList";
-        public static string profilesRegPath = @"Software\Microsoft\Windows NT\CurrentVersion\ProfileList";
+        
+        public static string profilesPath       = @"c:\Users\";
+        public static string root               = "HKEY_LOCAL_MACHINE";
+        public static string keyName            = @"ProfileList";
+        public static string profilesRegPath    = @"Software\Microsoft\Windows NT\CurrentVersion\ProfileList";
         public static List<Profile> profileList;
-        public static List<string> indexItems = new List<string>();
+        public static List<string> indexItems   = new List<string>();
  
         [STAThread]
         static void Main()
         {
+            
             profileList = RetrieveProfiles();
             foreach(Profile prof in profileList)
             {
                 indexItems.Add(prof.path);
             }
+
+
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
             Application.Run(new ProfileSelect());
@@ -36,9 +42,12 @@ namespace User_Cleanup
         private static List<Profile> RetrieveProfiles()
         {
 
-            string[] keys = Registry.LocalMachine.OpenSubKey(profilesRegPath).GetSubKeyNames();
+            string[] keys       = Registry.LocalMachine.OpenSubKey(profilesRegPath).GetSubKeyNames();
+            string[] subFolders = Directory.GetDirectories(@"c:\users");
+            
             List<Profile> results = new List<Profile>();
 
+            // REG RETRIEVE 
             foreach (var item in keys)
             {
                 Profile profile = new Profile();
@@ -46,7 +55,6 @@ namespace User_Cleanup
                 profile.registryKeyName = item;
                 try
                 {
-                    
                     // retrieve the key
                     profile.registryKey = Registry.LocalMachine.OpenSubKey(profilesRegPath).OpenSubKey(item);
 
@@ -60,17 +68,54 @@ namespace User_Cleanup
 
                     string[] arr = testval.ToString().Split('\\');
                     profile.name = arr[arr.Length - 1];
-
-
                     results.Add(profile);
 
                 } catch (System.NullReferenceException e)
                 {
 
-                }
-                
-
+                }    
             }
+
+            // Checks the profile directory
+            foreach (var item in subFolders)
+            {
+                // Value to determine if the profile is already in the list
+                bool cont = true;
+                try
+                {
+                    // iterates throguh profile list, checks the path value of each item
+                    foreach(Profile e in results)
+                    {
+                        if(e.path.ToLower().Equals(item.ToLower()))
+                        {
+                            // if a matching value is found, continue is set to false
+                            cont = false;
+                        }
+                    }
+
+
+                    // if continue remains true
+                    if (cont == true)
+                    {
+                        // a new profile item is created.
+                        var prof = new Profile();
+                        string[] arr = item.Split('\\');
+                        
+                        // Get the name off the path
+                        prof.path = item;
+                        prof.name = arr[arr.Length - 1];
+
+                        // add profile to list
+                        results.Add(prof);
+                    }
+
+                } catch (Exception _e)
+                {
+
+                }
+            }
+
+            // "Users" subfolder return. 
 
             return results;
         }
@@ -83,7 +128,8 @@ namespace User_Cleanup
             
             try
             {
-                Registry.LocalMachine.OpenSubKey(profilesRegPath,true).DeleteSubKey(userProfile.registryKeyName, false);
+                if(userProfile.registryKey != null)
+                    Registry.LocalMachine.OpenSubKey(profilesRegPath,true).DeleteSubKey(userProfile.registryKeyName, false);
             }
             catch (System.ArgumentException)
             {
@@ -95,50 +141,144 @@ namespace User_Cleanup
             }
 
             List<string> files = GetDirectoryContents(userProfile.path);
-
-            foreach (string file in files)
-            {
-                if (File.GetAttributes(file) == FileAttributes.Directory)
-                {
-                    Directory.Delete(file);
-                }
-                else
-                {
-                    File.Delete(file);
-                }
-
-            }
-
-        }
-
-        public static List<string> GetDirectoryContents(string path)
-        {
-            // Create a bar for getting directory contents(??)
             
-            List<string> files = new List<string>();
             try
             {
-                foreach(string file in Directory.GetFiles(path))
+                
+                foreach(string file in files)
                 {
-                    files.Add(file);
-                }
 
-                foreach(string directory in Directory.GetDirectories(path))
-                {
-                    files.AddRange(Directory.GetFiles(path));
                 }
-
-            }catch(Exception e)
+            
+            }
+            catch (Exception e)
             {
                 MessageBox.Show(e.Message, "Error!", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
+            
+
+        }
+
+        static List<string> GetDirectoryContents(string path)
+        {
+            List<string> files = new List<string>();
+            var fileSecurity = Directory.GetAccessControl(path);
+            fileSecurity.AddAccessRule(new FileSystemAccessRule("administrators", FileSystemRights.FullControl, InheritanceFlags.ContainerInherit,
+                PropagationFlags.InheritOnly, AccessControlType.Allow));
+            fileSecurity.AddAccessRule(new FileSystemAccessRule("administrators", FileSystemRights.FullControl, InheritanceFlags.ObjectInherit,
+                PropagationFlags.InheritOnly, AccessControlType.Allow));
+
+            // Get All Directories 
+            // Parent task
+            foreach (string directory in (IEnumerable<string>)Directory.EnumerateDirectories(path))
+            {
+                //printf(directory);
+                // If they are not in the files list, adds the range. 
+                if (!files.Contains(directory))
+                {
+                    files.Add(directory);
+                    files.AddRange(Directory.GetDirectories(directory));
+                }
+            }
+
+            // get the first object in files
+
+            List<string> updated = new List<string>();
+            updated.AddRange(files);
+
+
+
+            bool fin = false;
+
+            do
+            {
+                foreach (string subDir in files)
+                {
+                    fin = true;
+                    // Put that one object in a loop
+                    // Get the subdirectories of that one object and iterate until unable
+
+                    try
+                    {
+                        if (Directory.GetDirectories(subDir) != null)
+                        {
+                            foreach (string dir in Directory.GetDirectories(subDir))
+                            {
+                                if (!updated.Contains(dir))
+                                {
+                                    updated.Add(dir);
+                                    try
+                                    {
+                                        updated.AddRange(Directory.GetDirectories(dir));
+                                        fin = false;
+                                        printf($"ap@: {dir}");
+                                    }
+                                    catch (UnauthorizedAccessException _ex)
+                                    {
+                                        printf("setting file security");
+                                        Directory.SetAccessControl(dir, fileSecurity);
+                                        fin = false;
+                                    }
+                                    catch (FileNotFoundException _ex)
+                                    {
+                                        FileAttributes fa = File.GetAttributes(dir);
+                                        // Clear attributes
+                                        if (fa.HasFlag(FileAttributes.Hidden))
+                                        {
+                                            printf("removing hidden stuff");
+                                            var info = new DirectoryInfo(dir);
+                                            info.Attributes &= ~FileAttributes.Hidden;
+                                        }
+
+
+                                        if (fa.HasFlag(FileAttributes.ReadOnly))
+                                        {
+                                            printf("changing to not read only");
+                                            var info = new DirectoryInfo(dir);
+                                            info.Attributes &= ~FileAttributes.ReadOnly;
+                                        }
+                                    }
+
+                                }
+                            }
+                        }
+                    }
+                    catch (AccessViolationException _ex)
+                    {
+                        printf("setting file security");
+                        Directory.SetAccessControl(subDir, fileSecurity);
+                        fin = false;
+                    }
+
+
+                    /*
+                     * 
+                    if (!files.Contains(subDir)) 
+                    {
+                        files.AddRange(Directory.GetDirectories(subDir));
+                    }
+
+                    */
+
+                }
+                files = updated;
+            } while (!fin);
+            printf("files retrieved.");
 
             return files;
         }
 
+        static void printf(string str)
+        {
+            Console.WriteLine(str);
+        }
+
     }
 
-    class Profile
+
+}
+
+class Profile
     {
         public string name;
         public string path { get; set; }
